chnaged declaration

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,7 +10,7 @@
 
 int main(void)
 {
-	long long first = 0;
+	long first = 0;
 	int second = 1;
 	int next;
 	int i = 0;
